refactor(layout): migrate navigation and footer to widgets architecture
- Move Navigation components to widgets/menu structure
- Move Footer component to widgets/footer structure
- Update imports in layout files
- Convert function declarations to arrow functions
- Rename components for better semantics (Navbar → DesktopNavbar, MobileMenu → MobileNavbar)
