--- app/[language]/(guest)/(default-layout)/_layout/Footer/index.ts
@@ -1,1 +1,0 @@
-export { default as Footer } from './Footer';
