--- app/[language]/(guest)/(default-layout)/_layout/Navigation/index.ts
@@ -1,3 +1,0 @@
-export { Navbar } from './Navbar';
-export { MobileMenu } from './MobileMenu';
-export { FloatButtons } from './FloatButtons';
