--- app/[language]/(guest)/(default-layout)/_layout/index.ts
@@ -1,2 +1,0 @@
-export { FloatButtons, Navbar, MobileMenu } from './Navigation';
-export { Footer } from './Footer';
