+++ widgets/index.ts
@@ -1,0 +1,2 @@
+export * from './menu';
+export * from './footer';
