+++ widgets/menu/model/constants.ts
