+++ widgets/menu/model/index.ts
@@ -1,0 +1,1 @@
+export { NAV_ITEMS } from './constants';
